--- src/main/java/com/pabi/pabiuser/domain/Address/AddressGroup.java
@@ -1,5 +1,0 @@
-package com.pabi.pabiuser.domain.Address;
-
-public interface AddressGroup {
-
-}
