+++ src/main/java/com/pabi/pabiuser/domain/user/User.java
@@ -1,7 +1,6 @@
 package com.pabi.pabiuser.domain.user;
 
-import com.pabi.pabiuser.domain.Address.Address;
-import com.pabi.pabiuser.domain.user.UserCommand.Request;
+import com.pabi.pabiuser.domain.user.address.Address;
 import com.sun.istack.NotNull;
 import java.util.UUID;
 import javax.persistence.Column;
@@ -12,12 +11,9 @@ import javax.persistence.GenerationType;
 import javax.persistence.Id;
 import javax.persistence.Table;
 import lombok.AccessLevel;
-import lombok.Builder.Default;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
-import lombok.Value;
-import org.hibernate.annotations.ColumnDefault;
 import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
 
 @Entity
