+++ src/main/java/com/pabi/pabiuser/domain/user/UserInfo.java
@@ -1,6 +1,6 @@
 package com.pabi.pabiuser.domain.user;
 
-import com.pabi.pabiuser.domain.Address.Address;
+import com.pabi.pabiuser.domain.user.address.Address;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Getter;
@@ -11,6 +11,7 @@ public class UserInfo {
   @Builder
   @AllArgsConstructor
   public static class Main {
+ 
     private Long id;
     private String email;
     private String nickName;
