+++ src/main/java/com/pabi/pabiuser/domain/user/address/AddressGroup.java
@@ -1,0 +1,5 @@
+package com.pabi.pabiuser.domain.user.address;
+
+public interface AddressGroup {
+
+}
