+++ src/main/java/com/pabi/pabiuser/domain/user/address/NotNullAddress.java
@@ -1,4 +1,4 @@
-package com.pabi.pabiuser.domain.Address;
+package com.pabi.pabiuser.domain.user.address;
 
 import static java.lang.annotation.ElementType.TYPE;
 import static java.lang.annotation.RetentionPolicy.RUNTIME;
@@ -15,9 +15,9 @@ import javax.validation.Payload;
 @Documented
 public @interface NotNullAddress {
 
-    String message() default "";
+  String message() default "";
 
-    Class<?>[] groups() default {};
+  Class<?>[] groups() default {};
 
-    Class<? extends Payload>[] payload() default {};
+  Class<? extends Payload>[] payload() default {};
 }
